tray/requirements.txt
pystray>=0.19.0
pillow>=9.0.0
pyyaml>=6.0
